Update deploy script to run link checker
	modified:   deploy.sh

+++ deploy.sh
@@ -1,5 +1,12 @@
 #!/bin/bash
 
-rm -rf public/* && hugo && cd public && git add --all && git commit -m "Publishing to gh-pages"
+hugo serve &
 
-git push --all
+muffet --verbose http://localhost:1313/blog/
+exitCode=$?
+if [[ $exitCode == 0 ]]; then
+    rm -rf public/* && hugo && cd public && git add --all && git commit -m "Publishing to gh-pages"
+    git push --all
+fi
+
+kill $(ps aux | grep "hugo serve" | grep -v "grep" | awk '{print $2}')
